version 2/requirements.txt
# Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# Environment & Security
python-dotenv==1.0.0
bcrypt==4.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-flask==1.3.0